feat: implement import and export statements instead of IIFE
