+++ index.html
@@ -9,13 +9,6 @@
     <link rel="stylesheet" href="styles/main.css">
 </head>
 <body>
-    <script src="scripts/config.js"></script>
-    <script src="scripts/textArea.js"></script>
-    <script src="scripts/button.js"></script>
-    <script src="scripts/keyboard.js"></script>
-    <script src="scripts/keyboardListeners.js"></script>
-    <script src="scripts/infoText.js"></script>
-    <script src="scripts/helpers.js"></script>
-    <script src="scripts/main.js"></script>
+    <script src="scripts/main.js" type="module"></script>
 </body>
 </html>
