+++ scripts/config/config.js
@@ -1,0 +1,345 @@
+export const keyboardArray = [
+    {
+        code: 'Backquote',
+        ru: 'ё',
+        en: '`'
+    },
+    {
+        code: 'Digit1',
+        ru: '1',
+        en: '1'
+    },
+    {
+        code: 'Digit2',
+        ru: '2',
+        en: '2'
+    },
+    {
+        code: 'Digit3',
+        ru: '3',
+        en: '3'
+    },
+    {
+        code: 'Digit4',
+        ru: '4',
+        en: '4'
+    },
+    {
+        code: 'Digit5',
+        ru: '5',
+        en: '5'
+    },
+    {
+        code: 'Digit6',
+        ru: '6',
+        en: '6'
+    },
+    {
+        code: 'Digit7',
+        ru: '7',
+        en: '7'
+    },
+    {
+        code: 'Digit8',
+        ru: '8',
+        en: '8'
+    },
+    {
+        code: 'Digit9',
+        ru: '9',
+        en: '9'
+    },
+    {
+        code: 'Digit0',
+        ru: '0',
+        en: '0'
+    },
+    {
+        code: 'Minus',
+        ru: '-',
+        en: '-'
+    },
+    {
+        code: 'Equal',
+        ru: '=',
+        en: '='
+    },
+    {
+        code: 'Backspace',
+        label: 'Backspace',
+        isOperating: true,
+        additionalClasses: ['keyboard-button_name-backspace', 'keyboard-button_dark']
+    },
+    {
+        code: 'Tab',
+        label: 'Tab',
+        isOperating: true,
+        content: '\t',
+        additionalClasses: ['keyboard-button_dark']
+    },
+    {
+        code: 'KeyQ',
+        ru: 'й',
+        en: 'q'
+    },
+    {
+        code: 'KeyW',
+        ru: 'ц',
+        en: 'w'
+    },
+    {
+        code: 'KeyE',
+        ru: 'у',
+        en: 'e'
+    },
+    {
+        code: 'KeyR',
+        ru: 'к',
+        en: 'r'
+    },
+    {
+        code: 'KeyT',
+        ru: 'е',
+        en: 't'
+    },
+    {
+        code: 'KeyY',
+        ru: 'н',
+        en: 'y'
+    },
+    {
+        code: 'KeyU',
+        ru: 'г',
+        en: 'u'
+    },
+    {
+        code: 'KeyI',
+        ru: 'ш',
+        en: 'i'
+    },
+    {
+        code: 'KeyO',
+        ru: 'щ',
+        en: 'o'
+    },
+    {
+        code: 'KeyP',
+        ru: 'з',
+        en: 'p'
+    },
+    {
+        code: 'BracketLeft',
+        ru: 'х',
+        en: '['
+    },
+    {
+        code: 'BracketRight',
+        ru: 'ъ',
+        en: ']'
+    },
+    {
+        code: 'Backslash',
+        ru: '\\'
+    },
+    {
+        code: 'Delete',
+        label: 'Del',
+        isOperating: true,
+        additionalClasses: ['keyboard-button_dark']
+    },
+    {
+        code: 'CapsLock',
+        label: 'CapsLock',
+        isOperating: true,
+        additionalClasses: ['keyboard-button_dark', 'keyboard-button_name-capslock']
+    },
+    {
+        code: 'KeyA',
+        ru: 'ф',
+        en: 'a'
+    },
+    {
+        code: 'KeyS',
+        ru: 'ы',
+        en: 's'
+    },
+    {
+        code: 'KeyD',
+        ru: 'в',
+        en: 'd'
+    },
+    {
+        code: 'KeyF',
+        ru: 'а',
+        en: 'f'
+    },
+    {
+        code: 'KeyG',
+        ru: 'п',
+        en: 'g'
+    },
+    {
+        code: 'KeyH',
+        ru: 'р',
+        en: 'h'
+    },
+    {
+        code: 'KeyJ',
+        ru: 'о',
+        en: 'j'
+    },
+    {
+        code: 'KeyK',
+        ru: 'л',
+        en: 'k'
+    },
+    {
+        code: 'KeyL',
+        ru: 'д',
+        en: 'l'
+    },
+    {
+        code: 'Semicolon',
+        ru: 'ж',
+        en: ';'
+    },
+    {
+        code: 'Quote',
+        ru: 'э',
+        en: '\''
+    },
+    {
+        code: 'Enter',
+        label: 'Enter',
+        isOperating: true,
+        content: '\n',
+        additionalClasses: ['keyboard-button_name-enter', 'keyboard-button_dark']
+    },
+    {
+        code: 'ShiftLeft',
+        label: 'Shift',
+        isOperating: true,
+        additionalClasses: ['keyboard-button_name-shift-left', 'keyboard-button_dark']
+    },
+    {
+        code: 'KeyZ',
+        ru: 'я',
+        en: 'z'
+    },
+    {
+        code: 'KeyX',
+        ru: 'ч',
+        en: 'x'
+    },
+    {
+        code: 'KeyC',
+        ru: 'с',
+        en: 'c'
+    },
+    {
+        code: 'KeyV',
+        ru: 'м',
+        en: 'v'
+    },
+    {
+        code: 'KeyB',
+        ru: 'и',
+        en: 'b'
+    },
+    {
+        code: 'KeyN',
+        ru: 'т',
+        en: 'n'
+    },
+    {
+        code: 'KeyM',
+        ru: 'ь',
+        en: 'm'
+    },
+    {
+        code: 'Comma',
+        ru: 'б',
+        en: ','
+    },
+    {
+        code: 'Period',
+        ru: 'ю',
+        en: '.'
+    },
+    {
+        code: 'Slash',
+        ru: '.',
+        en: '/'
+    },
+    {
+        code: 'ArrowUp',
+        label: '<i class="fas fa-angle-up"></i>',
+        isOperating: true,
+        content: '▲',
+        additionalClasses: ['keyboard-button_dark']
+    },
+    {
+        code: 'ShiftRight',
+        label: 'Shift',
+        isOperating: true,
+        additionalClasses: ['keyboard-button_name-shift-right', 'keyboard-button_dark']
+    },
+    {
+        code: 'ControlLeft',
+        label: 'Ctrl',
+        isOperating: true,
+        additionalClasses: ['keyboard-button_dark']
+    },
+    {
+        code: 'MetaLeft',
+        label: 'Win',
+        isOperating: true,
+        additionalClasses: ['keyboard-button_dark']
+    },
+    {
+        code: 'AltLeft',
+        label: 'Alt',
+        isOperating: true,
+        additionalClasses: ['keyboard-button_dark']
+    },
+    {
+        code: 'Space',
+        label: '',
+        isOperating: true,
+        content: ' ',
+        additionalClasses: ['keyboard-button_name-space']
+    },
+    {
+        code: 'AltRight',
+        label: 'Alt',
+        isOperating: true,
+        additionalClasses: ['keyboard-button_dark']
+    },
+    {
+        code: 'ArrowLeft',
+        label: '<i class="fas fa-angle-left"></i>',
+        isOperating: true,
+        content: '◄',
+        additionalClasses: ['keyboard-button_dark']
+    },
+    {
+        code: 'ArrowDown',
+        label: '<i class="fas fa-angle-down"></i>',
+        isOperating: true,
+        content: '▼',
+        additionalClasses: ['keyboard-button_dark']
+    },
+    {
+        code: 'ArrowRight',
+        label: '<i class="fas fa-angle-right"></i>',
+        isOperating: true,
+        content: '►',
+        additionalClasses: ['keyboard-button_dark']
+    },
+    {
+        code: 'ControlRight',
+        label: 'Ctrl',
+        isOperating: true,
+        additionalClasses: ['keyboard-button_dark']
+    }
+];
