+++ scripts/main.js
@@ -1,20 +1,27 @@
+
+import TextArea from './components/TextArea.js';
+import Keyboard from './components/Keyboard.js';
+import InfoText from './components/InfoText.js';
+import { listenKeyboardButtonEvents } from './helpers/keyboardListeners.js';
+import { keyboardArray } from './config/config.js';
+
 window.onload = function() {
 
-    let infoText = new window.InfoText([
+    let infoText = new InfoText([
         'Разработано на Windows',
         'Сочетание клавиш для переключение языка - Ctrl + Alt'
     ]);
     infoText.mount();
 
-    let textArea = new window.TextArea('textarea', 'textarea');
+    let textArea = new TextArea('textarea', 'textarea');
     textArea.mount();
 
-    let keyboard = new window.Keyboard(window.keyboardArray, {
+    let keyboard = new Keyboard(keyboardArray, {
         className: 'keyboard'
     });
     keyboard.generateButtonsArray();
     keyboard.generateTemplate(keyboard.buttons);
     keyboard.mount();
 
-    window.listenKeyboardButtonEvents(keyboard.toggleLang, keyboard.activateButton, keyboard.deactivateButton, keyboard.handlePhysicalKeyboardInput);
+    listenKeyboardButtonEvents(keyboard.toggleLang, keyboard.activateButton, keyboard.deactivateButton, keyboard.handlePhysicalKeyboardInput);
 }
